Laravel cashier

+++ database/migrations/2021_02_16_083639_add_price_to_books_table.php
@@ -14,7 +14,7 @@ class AddPriceToBooksTable extends Migration
     public function up()
     {
         Schema::table('books', function (Blueprint $table) {
-            $table->float('price')->after('description');
+            $table->integer('price')->after('description');
             $table->integer('discount')->after('price')->nullable();
         });
     }
